Highlight card everywhere when hovered
When hovering a card in your hand or in a presentation option, any place
that the card appears on the page will be highlighted. This is to make
it easier to understand where in your hand a presentation option occurs,
where to find the presentation option for a card in your hand, or which
presentation options share the same card.